fix crash

+++ core/src/main/java/net/labymod/addons/damageindicator/tags/HealthAmountTag.java
@@ -52,7 +52,7 @@ public class HealthAmountTag extends NameTag {
     super.render(stack, entity);
     RenderPipeline renderPipeline = this.addon.labyAPI().renderPipeline();
     LivingEntity livingEntity = (LivingEntity) entity;
-    int startX = renderPipeline.componentRenderer().width(this.getComponent(livingEntity)) + 2;
+    float startX = renderPipeline.componentRenderer().width(this.getComponent(livingEntity)) + 2;
     renderPipeline.renderSeeThrough(entity,
         () -> renderPipeline.resourceRenderer()
             .entityHeartRenderer(livingEntity).renderHealthBar(
